feat: add embedded ARM & Akai MPC Live 1 multi-target architecture specification
- Document ARMv7 target (`armv7-unknown-linux-gnueabihf`), 32-bit type system safety, and composable repository target layout in `docs/system/EMBEDDED_ARM_MPC_ARCHITECTURE.md`
- Harden `crates/ipc-layer/src/lib.rs` for 32-bit `rlim_cur` to `u64` conversions
- Update `docs/README.md` hub index

+++ crates/ipc-layer/src/lib.rs
@@ -511,7 +511,7 @@ impl std::fmt::Display for SchedStatus {
 pub fn rtprio_limit() -> Option<u64> {
     let mut lim = libc::rlimit { rlim_cur: 0, rlim_max: 0 };
     if unsafe { libc::getrlimit(libc::RLIMIT_RTPRIO, &mut lim) } == 0 {
-        Some(lim.rlim_cur)
+        Some(lim.rlim_cur as u64)
     } else {
         None
     }
@@ -1127,7 +1127,7 @@ pub fn memlock_limit() -> Option<u64> {
     {
         let mut lim = libc::rlimit { rlim_cur: 0, rlim_max: 0 };
         if unsafe { libc::getrlimit(libc::RLIMIT_MEMLOCK, &mut lim) } == 0 {
-            Some(lim.rlim_cur)
+            Some(lim.rlim_cur as u64)
         } else {
             None
         }
